# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UiRealtyPals API",
    "description": "Full REST API for UiRealtyPals - AI-powered real estate discovery platform for Noida. Auth: users send Authorization: Bearer <supabase-access-token>. Admin uses admin_session cookie. Webhooks use X-Webhook-Secret header.",
    "version": "1.0.0",
    "contact": {
      "name": "RealtyPals Engineering",
      "url": "https://realty-pals-dev-frontend-three.vercel.app"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev server"
    },
    {
      "url": "https://realtypalsdev.onrender.com",
      "description": "Production (Render)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token. Obtain via supabase.auth.signIn*(). Send as: Authorization: Bearer <token>"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session",
        "description": "Admin session cookie set by POST /api/v1/admin/login. Required for all /admin/* routes."
      },
      "WebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "HMAC-SHA256 secret for webhook lead ingestion."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Not found"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "UnitType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "bhk": {
            "type": "integer",
            "example": 3
          },
          "price_min_cr": {
            "type": "number",
            "example": 1.85,
            "description": "Minimum price in Crore INR"
          },
          "price_max_cr": {
            "type": "number",
            "example": 3.14
          },
          "super_area_sqft": {
            "type": "integer",
            "example": 1650
          },
          "carpet_area_sqft": {
            "type": "integer",
            "example": 1250
          }
        }
      },
      "ProjectImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "type": {
            "type": "string",
            "enum": [
              "hero",
              "gallery",
              "floor_plan",
              "amenity"
            ]
          },
          "sort_order": {
            "type": "integer"
          }
        }
      },
...
      "ProjectCard": {
        "type": "object",
        "description": "Lightweight project card for listing and chat responses",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Godrej Meridien"
          },
          "slug": {
            "type": "string",
            "example": "godrej-meridien-sector-106"
          },
          "tagline": {
            "type": "string"
          },
          "sector": {
            "type": "string",
            "example": "Sector 106"
          },
          "city": {
            "type": "string",
            "example": "Noida"
          },
          "status": {
            "type": "string",
            "enum": [
              "under_construction",
              "ready_to_move",
              "new_launch"
            ]
          },
          "possession_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "rera_number": {
            "type": "string",
            "nullable": true
          },
          "total_units": {
            "type": "integer",
            "nullable": true
          },
          "total_towers": {
            "type": "integer",
            "nullable": true
          },
          "land_area_acres": {
            "type": "number",
            "nullable": true
          },
          "builder": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              }
            }
          },
          "unit_types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitType"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectImage"
            }
          }
        }
      },
...
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - valid Bearer token or admin session required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Invalid request body or query parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/saved": {
      "get": {
        "tags": [
          "Saved"
        ],
        "summary": "Get saved properties",
        "description": "All saved properties for authenticated user ordered by most recently saved.",
        "operationId": "getSavedProperties",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Saved properties",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectCard"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Saved"
        ],
        "summary": "Save a property",
        "description": "Idempotent upsert. Also logs a PropertyEvent for analytics.",
        "operationId": "saveProperty",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "project_id"
                ],
                "properties": {
                  "project_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/saved/{id}/check": {
      "get": {
        "tags": [
          "Saved"
        ],
        "summary": "Check if property is saved",
        "description": ":id = project_id UUID",
        "operationId": "checkPropertySaved",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Save status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "is_saved": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/saved/{id}": {
      "delete": {
        "tags": [
          "Saved"
        ],
        "summary": "Remove saved property",
        "description": ":id = project_id. Logs PropertyEvent.",
        "operationId": "removeSavedProperty",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}